make BatchTiledMapRenderer extensions call original protected methods
