+++ tiled/src/main/kotlin/ktx/tiled/batchTiledMapRenderer.kt
@@ -1,74 +1,16 @@
-package ktx.tiled
+@file:Suppress("PackageDirectoryMismatch")
 
-import com.badlogic.gdx.graphics.OrthographicCamera
-import com.badlogic.gdx.maps.tiled.renderers.BatchTiledMapRenderer
-import com.badlogic.gdx.maps.tiled.tiles.AnimatedTiledMapTile
-import com.badlogic.gdx.math.Matrix4
-import com.badlogic.gdx.math.Rectangle
+package com.badlogic.gdx.maps.tiled.renderers
 
 /**
- * Automatically calls [BatchTiledMapRenderer.beginRender] and [BatchTiledMapRenderer.endRender].
- * @param camera A camera to set on the renderer before [BatchTiledMapRenderer.beginRender].
- * @param block inlined. Executed after [BatchTiledMapRenderer.beginRender] and before [BatchTiledMapRenderer.endRender].
+ * This file is used as a workaround for the tiledMapRenderer extensions. They need
+ * to call [BatchTiledMapRenderer.beginRender] and [BatchTiledMapRenderer.endRender] methods
+ * which are protected methods. Since this file matches the package of the [BatchTiledMapRenderer],
+ * we can call protected methods of it and use our wrapper functions for public API extensions.
  */
-inline fun <T : BatchTiledMapRenderer> T.use(
-  camera: OrthographicCamera,
-  block: (T) -> Unit
-) {
-  this.setView(camera)
-  AnimatedTiledMapTile.updateAnimationBaseTime()
-  this.batch.begin()
 
-  block(this)
+@PublishedApi
+internal fun BatchTiledMapRenderer.beginInternal() = this.beginRender()
 
-  this.batch.end()
-}
-
-/**
- * Automatically calls [BatchTiledMapRenderer.beginRender] and [BatchTiledMapRenderer.endRender].
- * @param projection A projection matrix to set on the renderer before [BatchTiledMapRenderer.beginRender].
- * @param x map render boundary x value.
- * @param y map render boundary y value.
- * @param width map render boundary width value.
- * @param height map render boundary height value.
- * @param block inlined. Executed after [BatchTiledMapRenderer.beginRender] and before [BatchTiledMapRenderer.endRender].
- */
-inline fun <T : BatchTiledMapRenderer> T.use(
-  projection: Matrix4,
-  x: Float, y: Float,
-  width: Float, height: Float,
-  block: (T) -> Unit
-) {
-  this.setView(projection, x, y, width, height)
-  AnimatedTiledMapTile.updateAnimationBaseTime()
-  this.batch.begin()
-
-  block(this)
-
-  this.batch.end()
-}
-
-/**
- * Automatically calls [BatchTiledMapRenderer.beginRender] and [BatchTiledMapRenderer.endRender].
- * @param projection A projection matrix to set on the renderer before [BatchTiledMapRenderer.beginRender].
- * @param mapBoundary map render boundary.
- * @param block inlined. Executed after [BatchTiledMapRenderer.beginRender] and before [BatchTiledMapRenderer.endRender].
- */
-inline fun <T : BatchTiledMapRenderer> T.use(
-  projection: Matrix4,
-  mapBoundary: Rectangle,
-  block: (T) -> Unit
-) = this.use(projection, mapBoundary.x, mapBoundary.y, mapBoundary.width, mapBoundary.height, block)
-
-/**
- * Automatically calls [BatchTiledMapRenderer.beginRender] and [BatchTiledMapRenderer.endRender].
- * @param block inlined. Executed after [BatchTiledMapRenderer.beginRender] and before [BatchTiledMapRenderer.endRender].
- */
-inline fun <T : BatchTiledMapRenderer> T.use(block: (T) -> Unit) {
-  AnimatedTiledMapTile.updateAnimationBaseTime()
-  this.batch.begin()
-
-  block(this)
-
-  this.batch.end()
-}
+@PublishedApi
+internal fun BatchTiledMapRenderer.endInternal() = this.endRender()
